backend_checker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dacite==1.8.1
exceptiongroup==1.2.1
fastapi==0.110.3
h11==0.14.0
httptools==0.6.1
idna==3.7
inject==5.2.1
mashumaro==3.13
nats-py==2.4.0
orjson==3.10.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
